+++ src/board/piece_movement.rs
@@ -152,15 +152,15 @@ impl Board {
         todo!()
     }
     pub fn put_piece(&mut self, piece: Piece, side: Side, square_idx: Square) {
-        self.bb_pieces[side as usize][piece] |= SQUARE_MASKS[square_idx];
-        self.bb_sides[side as usize] |= SQUARE_MASKS[square_idx];
+        self.bb_pieces[side][piece] |= SQUARE_MASKS[square_idx];
+        self.bb_sides[side] |= SQUARE_MASKS[square_idx];
         self.piece_list[square_idx] = piece;
         self.game_state.zobrist_key ^= self.zobrist_hashmap.piece(side, piece, square_idx)
     }
 
     pub fn remove_piece(&mut self, piece: Piece, side: Side, square_idx: Square) {
-        self.bb_pieces[side as usize][piece] &= !SQUARE_MASKS[square_idx];
-        self.bb_sides[side as usize] &= !SQUARE_MASKS[square_idx];
+        self.bb_pieces[side][piece] &= !SQUARE_MASKS[square_idx];
+        self.bb_sides[side] &= !SQUARE_MASKS[square_idx];
         self.piece_list[square_idx] = Pieces::NONE;
         self.game_state.zobrist_key ^= self.zobrist_hashmap.piece(side, piece, square_idx);
     }
